refactor(dgw): move the session store out of the credential module
The store keyed by association-token JTI is not credential-specific: a session can
need other things provisioned ahead of the connection, and the next commit adds the
first one. Move it to a `provisioning` module and name it for what it holds, leaving
`credential` with just the credential types.

Pure move and rename — same types, same API, same behavior. Callers follow.
