+++ devolutions-gateway/src/credential/mod.rs
@@ -3,40 +3,9 @@ mod crypto;
 #[rustfmt::skip]
 pub use crypto::EncryptedPassword;
 
-use std::collections::HashMap;
-use std::fmt;
-use std::sync::Arc;
-
-use anyhow::Context;
-use async_trait::async_trait;
-use devolutions_gateway_task::{ShutdownSignal, Task};
-use parking_lot::Mutex;
 use secrecy::ExposeSecret as _;
-use uuid::Uuid;
 
 use self::crypto::MASTER_KEY;
-
-/// Error returned by [`CredentialStoreHandle::insert`].
-#[derive(Debug)]
-pub enum InsertError {
-    /// The provided token is invalid (e.g., missing or malformed JTI).
-    ///
-    /// This is a client-side error: the caller supplied bad input.
-    InvalidToken(anyhow::Error),
-    /// An internal error occurred (e.g., encryption failure).
-    Internal(anyhow::Error),
-}
-
-impl fmt::Display for InsertError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::InvalidToken(e) => e.fmt(f),
-            Self::Internal(e) => e.fmt(f),
-        }
-    }
-}
-
-impl std::error::Error for InsertError {}
 
 /// Credential at the application protocol level
 #[derive(Debug, Clone)]
@@ -70,8 +39,8 @@ pub struct AppCredentialMapping {
 
 /// Cleartext credential received from the API, used for deserialization only.
 ///
-/// Passwords are encrypted and stored as [`AppCredential`] inside the credential store.
-/// This type is never stored directly — hand it to [`CredentialStoreHandle::insert`].
+/// Passwords are encrypted and stored as [`AppCredential`] inside the provisioning store.
+/// This type is never stored directly — hand it to [`ProvisioningStore::insert`].
 #[derive(Debug, Deserialize)]
 #[serde(tag = "kind")]
 pub enum CleartextAppCredential {
@@ -98,7 +67,7 @@ impl CleartextAppCredential {
 
 /// Cleartext credential mapping received from the API, used for deserialization only.
 ///
-/// Passwords are encrypted on write. Hand this directly to [`CredentialStoreHandle::insert`].
+/// Passwords are encrypted on write. Hand this directly to [`ProvisioningStore::insert`].
 #[derive(Debug, Deserialize)]
 pub struct CleartextAppCredentialMapping {
     #[serde(rename = "proxy_credential")]
@@ -108,129 +77,10 @@ pub struct CleartextAppCredentialMapping {
 }
 
 impl CleartextAppCredentialMapping {
-    fn encrypt(self) -> anyhow::Result<AppCredentialMapping> {
+    pub(crate) fn encrypt(self) -> anyhow::Result<AppCredentialMapping> {
         Ok(AppCredentialMapping {
             proxy: self.proxy.encrypt()?,
             target: self.target.encrypt()?,
         })
     }
-}
-
-#[derive(Debug, Clone)]
-pub struct CredentialStoreHandle(Arc<Mutex<CredentialStore>>);
-
-impl Default for CredentialStoreHandle {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl CredentialStoreHandle {
-    pub fn new() -> Self {
-        Self(Arc::new(Mutex::new(CredentialStore::new())))
-    }
-
-    pub fn insert(
-        &self,
-        token: String,
-        mapping: Option<CleartextAppCredentialMapping>,
-        time_to_live: time::Duration,
-    ) -> Result<Option<ArcCredentialEntry>, InsertError> {
-        let mapping = mapping
-            .map(CleartextAppCredentialMapping::encrypt)
-            .transpose()
-            .map_err(InsertError::Internal)?;
-        self.0.lock().insert(token, mapping, time_to_live)
-    }
-
-    pub fn get(&self, token_id: Uuid) -> Option<ArcCredentialEntry> {
-        self.0.lock().get(token_id)
-    }
-}
-
-#[derive(Debug)]
-struct CredentialStore {
-    entries: HashMap<Uuid, ArcCredentialEntry>,
-}
-
-#[derive(Debug)]
-pub struct CredentialEntry {
-    pub token: String,
-    pub mapping: Option<AppCredentialMapping>,
-    pub expires_at: time::OffsetDateTime,
-}
-
-pub type ArcCredentialEntry = Arc<CredentialEntry>;
-
-impl CredentialStore {
-    fn new() -> Self {
-        Self {
-            entries: HashMap::new(),
-        }
-    }
-
-    fn insert(
-        &mut self,
-        token: String,
-        mapping: Option<AppCredentialMapping>,
-        time_to_live: time::Duration,
-    ) -> Result<Option<ArcCredentialEntry>, InsertError> {
-        let jti = crate::token::extract_jti(&token)
-            .context("failed to extract token ID")
-            .map_err(InsertError::InvalidToken)?;
-
-        let entry = CredentialEntry {
-            token,
-            mapping,
-            expires_at: time::OffsetDateTime::now_utc() + time_to_live,
-        };
-
-        let previous_entry = self.entries.insert(jti, Arc::new(entry));
-
-        Ok(previous_entry)
-    }
-
-    fn get(&self, token_id: Uuid) -> Option<ArcCredentialEntry> {
-        self.entries.get(&token_id).map(Arc::clone)
-    }
-}
-
-pub struct CleanupTask {
-    pub handle: CredentialStoreHandle,
-}
-
-#[async_trait]
-impl Task for CleanupTask {
-    type Output = anyhow::Result<()>;
-
-    const NAME: &'static str = "credential store cleanup";
-
-    async fn run(self, shutdown_signal: ShutdownSignal) -> Self::Output {
-        cleanup_task(self.handle, shutdown_signal).await;
-        Ok(())
-    }
-}
-
-#[instrument(skip_all)]
-async fn cleanup_task(handle: CredentialStoreHandle, mut shutdown_signal: ShutdownSignal) {
-    use tokio::time::{Duration, sleep};
-
-    const TASK_INTERVAL: Duration = Duration::from_secs(60 * 15); // 15 minutes
-
-    debug!("Task started");
-
-    loop {
-        tokio::select! {
-            _ = sleep(TASK_INTERVAL) => {}
-            _ = shutdown_signal.wait() => {
-                break;
-            }
-        }
-
-        let now = time::OffsetDateTime::now_utc();
-
-        handle.0.lock().entries.retain(|_, src| now < src.expires_at);
-    }
-
-    debug!("Task terminated");
 }
